game1つに対し、複数のplatformを表示できるように変更（途中）
