+++ src/main/java/com/crud_read_and_create/entity/Game.java
@@ -1,19 +1,20 @@
 package com.crud_read_and_create.entity;
 
+import java.util.List;
+
 public class Game {
 	private String id;
 	private String name;
 	private String genre;
-	private String platform;
 	private Integer price;
+	private List<Platform> platforms;
 
-	public Game(String id, String name, String genre, String platform, Integer price) {
+	public Game(String id, String name, String genre, Integer price, List<Platform> platforms) {
 		this.id = id;
 		this.name = name;
 		this.genre = genre;
-		this.platform = platform;
 		this.price = price;
-
+		this.platforms = platforms;
 	}
 
 	public Game() {
@@ -24,40 +25,40 @@ public class Game {
 		return id;
 	}
 
-	public void setId(String id) {
-		this.id = id;
-	}
-
 	public String getName() {
 		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
 	}
 
 	public String getGenre() {
 		return genre;
 	}
 
-	public void setGenre(String genre) {
-		this.genre = genre;
-	}
-
-	public String getPlatform() {
-		return platform;
-	}
-
-	public void setPlatform(String platform) {
-		this.platform = platform;
-	}
-
 	public Integer getPrice() {
 		return price;
 	}
 
+	public List<Platform> getPlatforms() {
+		return platforms;
+	}
+
+	public void setId(String id) {
+		this.id = id;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	public void setGenre(String genre) {
+		this.genre = genre;
+	}
+
 	public void setPrice(Integer price) {
 		this.price = price;
+	}
+
+	public void setPlatforms(List<Platform> platforms) {
+		this.platforms = platforms;
 	}
 
 }
