+++ src/main/java/com/crud_read_and_create/entity/Platform.java
@@ -1,0 +1,22 @@
+package com.crud_read_and_create.entity;
+
+public class Platform {
+	private String platform;
+
+	public Platform(String platform) {
+		this.platform = platform;
+	}
+
+	public Platform() {
+
+	}
+
+	public String getPlatform() {
+		return platform;
+	}
+
+	public void setPlatform(String platform) {
+		this.platform = platform;
+	}
+
+}
