+++ src/main/resources/com/crud_read_and_create/mapper/GameMapper.xml
@@ -3,35 +3,48 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">   
 <mapper namespace="com.crud_read_and_create.mapper.GameMapper">
-    <select id="findAllAsc" resultType="com.crud_read_and_create.entity.Game">
+	
+	<resultMap type="com.crud_read_and_create.entity.Game" id="gameMap">
+		<id property="games.id" column="games.id"/>
+		<result property="id" column="id" />
+		<result property="name" column="name" />
+		<result property="genre" column="genre" />
+		<result property="price" column="price" />
+		
+		<collection property="platforms" ofType="com.crud_read_and_create.entity.Platform">
+			<id property="platforms.id" column="platforms.id"/>
+			<result property="platform" column="platform"/>		
+		</collection>
+	</resultMap>
+	
+    <select id="findAllAsc" resultMap="gameMap">
         SELECT games.id, name, platform, genre, price 
-        FROM games_platforms
-        INNER JOIN games
- 		ON games_platforms.games_id = games.id
+        FROM games
+        INNER JOIN games_platforms
+ 		ON games.id = games_platforms.games_id
  		INNER JOIN platforms
  		ON games_platforms.platforms_id = platforms.id
- 		ORDER BY games.id ASC;
-        
+ 		ORDER BY games_platforms.id ASC;
     </select>
     
-    <select id="findAllDesc" resultType="com.crud_read_and_create.entity.Game">
+    <select id="findAllDesc" resultMap="gameMap">
         SELECT games.id, name, platform, genre, price 
-        FROM games_platforms
-        INNER JOIN games
- 		ON games_platforms.games_id = games.id
+        FROM games
+        INNER JOIN games_platforms
+ 		ON games.id = games_platforms.games_id
  		INNER JOIN platforms
  		ON games_platforms.platforms_id = platforms.id
- 		ORDER BY games.id DESC;
+ 		ORDER BY games_platforms.id DESC;
     </select>
     
-    <select id="findById" resultType="com.crud_read_and_create.entity.Game">
+    <select id="findById" resultMap="gameMap">
         SELECT games.id, name, platform, genre, price 
         FROM games_platforms
         INNER JOIN games
  		ON games_platforms.games_id = games.id
  		INNER JOIN platforms
  		ON games_platforms.platforms_id = platforms.id
-        WHERE games.id = #{id} 
+        WHERE games_platforms.id = #{id} 
     </select>
     
     <insert id="create" useGeneratedKeys="true" keyProperty="id">
