+++ src/main/resources/templates/search/db.html
@@ -48,7 +48,7 @@
             		<td th:text="*{id}">ID</td>
             		<td th:text="*{name}">タイトル</td>
             		<td th:text="*{genre}">ジャンル</td>
-            		<td th:text="*{platform}">プラットフォーム</td>
+            		<td th:text="*{platform}">プラットフォーム	</td>
             		<td th:text="*{price}">価格</td>
         		</tr>  
     		</table>
